requirements.txt
requests
Pillow
inky[rpi]
